% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_varcheck}
\alias{theme_varcheck}
\title{Default VARcheck ggplot2 theme}
\usage{
theme_varcheck()
}
\value{
A `ggplot2` theme object.
}
\description{
A minimal theme used as the default style for all VARcheck plots.
Pass a `ggplot2::theme()` object to the `theme` argument of
[plot_var_check()] to override individual elements on top of this base.
}
\examples{
theme_varcheck()
}
